refactor(frontends/basic): split Lowerer into submodules
