+++ src/frontends/basic/Lowerer.cpp
@@ -84,547 +84,6 @@ Module Lowerer::lower(const Program &prog)
     return lowerProgram(prog);
 }
 
-// Purpose: declare required runtime.
-// Parameters: build::IRBuilder &b.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::declareRequiredRuntime(build::IRBuilder &b)
-{
-    using Type = il::core::Type;
-    b.addExtern("rt_print_str", Type(Type::Kind::Void), {Type(Type::Kind::Str)});
-    b.addExtern("rt_print_i64", Type(Type::Kind::Void), {Type(Type::Kind::I64)});
-    b.addExtern("rt_print_f64", Type(Type::Kind::Void), {Type(Type::Kind::F64)});
-    b.addExtern("rt_len", Type(Type::Kind::I64), {Type(Type::Kind::Str)});
-    b.addExtern("rt_substr",
-                Type(Type::Kind::Str),
-                {Type(Type::Kind::Str), Type(Type::Kind::I64), Type(Type::Kind::I64)});
-    if (needRtConcat)
-        b.addExtern(
-            "rt_concat", Type(Type::Kind::Str), {Type(Type::Kind::Str), Type(Type::Kind::Str)});
-    if (boundsChecks)
-        b.addExtern("rt_trap", Type(Type::Kind::Void), {Type(Type::Kind::Str)});
-    if (needInputLine)
-        b.addExtern("rt_input_line", Type(Type::Kind::Str), {});
-    if (needRtToInt)
-        b.addExtern("rt_to_int", Type(Type::Kind::I64), {Type(Type::Kind::Str)});
-    if (needRtIntToStr)
-        b.addExtern("rt_int_to_str", Type(Type::Kind::Str), {Type(Type::Kind::I64)});
-    if (needRtF64ToStr)
-        b.addExtern("rt_f64_to_str", Type(Type::Kind::Str), {Type(Type::Kind::F64)});
-    if (needAlloc)
-        b.addExtern("rt_alloc", Type(Type::Kind::Ptr), {Type(Type::Kind::I64)});
-    if (needRtLeft)
-        b.addExtern(
-            "rt_left", Type(Type::Kind::Str), {Type(Type::Kind::Str), Type(Type::Kind::I64)});
-    if (needRtRight)
-        b.addExtern(
-            "rt_right", Type(Type::Kind::Str), {Type(Type::Kind::Str), Type(Type::Kind::I64)});
-    if (needRtMid2)
-        b.addExtern(
-            "rt_mid2", Type(Type::Kind::Str), {Type(Type::Kind::Str), Type(Type::Kind::I64)});
-    if (needRtMid3)
-        b.addExtern("rt_mid3",
-                    Type(Type::Kind::Str),
-                    {Type(Type::Kind::Str), Type(Type::Kind::I64), Type(Type::Kind::I64)});
-    if (needRtInstr2)
-        b.addExtern(
-            "rt_instr2", Type(Type::Kind::I64), {Type(Type::Kind::Str), Type(Type::Kind::Str)});
-    if (needRtInstr3)
-        b.addExtern("rt_instr3",
-                    Type(Type::Kind::I64),
-                    {Type(Type::Kind::I64), Type(Type::Kind::Str), Type(Type::Kind::Str)});
-    if (needRtLtrim)
-        b.addExtern("rt_ltrim", Type(Type::Kind::Str), {Type(Type::Kind::Str)});
-    if (needRtRtrim)
-        b.addExtern("rt_rtrim", Type(Type::Kind::Str), {Type(Type::Kind::Str)});
-    if (needRtTrim)
-        b.addExtern("rt_trim", Type(Type::Kind::Str), {Type(Type::Kind::Str)});
-    if (needRtUcase)
-        b.addExtern("rt_ucase", Type(Type::Kind::Str), {Type(Type::Kind::Str)});
-    if (needRtLcase)
-        b.addExtern("rt_lcase", Type(Type::Kind::Str), {Type(Type::Kind::Str)});
-    if (needRtChr)
-        b.addExtern("rt_chr", Type(Type::Kind::Str), {Type(Type::Kind::I64)});
-    if (needRtAsc)
-        b.addExtern("rt_asc", Type(Type::Kind::I64), {Type(Type::Kind::Str)});
-
-    for (RuntimeFn fn : runtimeOrder)
-    {
-        switch (fn)
-        {
-            case RuntimeFn::Sqrt:
-                b.addExtern("rt_sqrt", Type(Type::Kind::F64), {Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::AbsI64:
-                b.addExtern("rt_abs_i64", Type(Type::Kind::I64), {Type(Type::Kind::I64)});
-                break;
-            case RuntimeFn::AbsF64:
-                b.addExtern("rt_abs_f64", Type(Type::Kind::F64), {Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::Floor:
-                b.addExtern("rt_floor", Type(Type::Kind::F64), {Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::Ceil:
-                b.addExtern("rt_ceil", Type(Type::Kind::F64), {Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::Sin:
-                b.addExtern("rt_sin", Type(Type::Kind::F64), {Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::Cos:
-                b.addExtern("rt_cos", Type(Type::Kind::F64), {Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::Pow:
-                b.addExtern("rt_pow",
-                            Type(Type::Kind::F64),
-                            {Type(Type::Kind::F64), Type(Type::Kind::F64)});
-                break;
-            case RuntimeFn::RandomizeI64:
-                b.addExtern("rt_randomize_i64", Type(Type::Kind::Void), {Type(Type::Kind::I64)});
-                break;
-            case RuntimeFn::Rnd:
-                b.addExtern("rt_rnd", Type(Type::Kind::F64), {});
-                break;
-        }
-    }
-
-    if (needRtStrEq)
-        b.addExtern(
-            "rt_str_eq", Type(Type::Kind::I1), {Type(Type::Kind::Str), Type(Type::Kind::Str)});
-}
-
-// Purpose: track runtime.
-// Parameters: RuntimeFn fn.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::trackRuntime(RuntimeFn fn)
-{
-    if (runtimeSet.insert(fn).second)
-        runtimeOrder.push_back(fn);
-}
-
-// Purpose: scan unary expr.
-// Parameters: const UnaryExpr &u.
-// Returns: Lowerer::ExprType.
-// Side effects: may modify lowering state or emit IL.
-Lowerer::ExprType Lowerer::scanUnaryExpr(const UnaryExpr &u)
-{
-    return scanExpr(*u.expr);
-}
-
-// Purpose: scan binary expr.
-// Parameters: const BinaryExpr &b.
-// Returns: Lowerer::ExprType.
-// Side effects: may modify lowering state or emit IL.
-Lowerer::ExprType Lowerer::scanBinaryExpr(const BinaryExpr &b)
-{
-    ExprType lt = scanExpr(*b.lhs);
-    ExprType rt = scanExpr(*b.rhs);
-    if (b.op == BinaryExpr::Op::Add && lt == ExprType::Str && rt == ExprType::Str)
-    {
-        needRtConcat = true;
-        return ExprType::Str;
-    }
-    if (b.op == BinaryExpr::Op::Eq || b.op == BinaryExpr::Op::Ne)
-    {
-        if (lt == ExprType::Str || rt == ExprType::Str)
-            needRtStrEq = true;
-        return ExprType::Bool;
-    }
-    if (lt == ExprType::F64 || rt == ExprType::F64)
-        return ExprType::F64;
-    return ExprType::I64;
-}
-
-// Purpose: scan array expr.
-// Parameters: const ArrayExpr &arr.
-// Returns: Lowerer::ExprType.
-// Side effects: may modify lowering state or emit IL.
-Lowerer::ExprType Lowerer::scanArrayExpr(const ArrayExpr &arr)
-{
-    scanExpr(*arr.index);
-    return ExprType::I64;
-}
-
-// Purpose: scan builtin call expr.
-// Parameters: const BuiltinCallExpr &c.
-// Returns: Lowerer::ExprType.
-// Side effects: may modify lowering state or emit IL.
-Lowerer::ExprType Lowerer::scanBuiltinCallExpr(const BuiltinCallExpr &c)
-{
-    switch (c.builtin)
-    {
-        case BuiltinCallExpr::Builtin::Rnd:
-            trackRuntime(RuntimeFn::Rnd);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Val:
-            needRtToInt = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::I64;
-        case BuiltinCallExpr::Builtin::Str:
-            needRtIntToStr = true;
-            needRtF64ToStr = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Len:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::I64;
-        case BuiltinCallExpr::Builtin::Left:
-            needRtLeft = true;
-            for (auto &a : c.args)
-                if (a)
-                    scanExpr(*a);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Right:
-            needRtRight = true;
-            for (auto &a : c.args)
-                if (a)
-                    scanExpr(*a);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Mid:
-            if (c.args.size() >= 3 && c.args[2])
-                needRtMid3 = true;
-            else
-                needRtMid2 = true;
-            for (auto &a : c.args)
-                if (a)
-                    scanExpr(*a);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Instr:
-            if (c.args.size() >= 3 && c.args[0])
-                needRtInstr3 = true;
-            else
-                needRtInstr2 = true;
-            for (auto &a : c.args)
-                if (a)
-                    scanExpr(*a);
-            return ExprType::I64;
-        case BuiltinCallExpr::Builtin::Ltrim:
-            needRtLtrim = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Rtrim:
-            needRtRtrim = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Trim:
-            needRtTrim = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Ucase:
-            needRtUcase = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Lcase:
-            needRtLcase = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Chr:
-            needRtChr = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::Str;
-        case BuiltinCallExpr::Builtin::Asc:
-            needRtAsc = true;
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::I64;
-        case BuiltinCallExpr::Builtin::Sqr:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            trackRuntime(RuntimeFn::Sqrt);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Abs:
-        {
-            ExprType ty = ExprType::I64;
-            if (c.args[0])
-                ty = scanExpr(*c.args[0]);
-            if (ty == ExprType::F64)
-                trackRuntime(RuntimeFn::AbsF64);
-            else
-                trackRuntime(RuntimeFn::AbsI64);
-            return ty;
-        }
-        case BuiltinCallExpr::Builtin::Floor:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            trackRuntime(RuntimeFn::Floor);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Ceil:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            trackRuntime(RuntimeFn::Ceil);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Sin:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            trackRuntime(RuntimeFn::Sin);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Cos:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            trackRuntime(RuntimeFn::Cos);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Pow:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            if (c.args[1])
-                scanExpr(*c.args[1]);
-            trackRuntime(RuntimeFn::Pow);
-            return ExprType::F64;
-        case BuiltinCallExpr::Builtin::Int:
-            if (c.args[0])
-                scanExpr(*c.args[0]);
-            return ExprType::I64;
-        default:
-            for (auto &a : c.args)
-                if (a)
-                    scanExpr(*a);
-            return ExprType::I64;
-    }
-}
-
-// Purpose: scan expr.
-// Parameters: const Expr &e.
-// Returns: Lowerer::ExprType.
-// Side effects: may modify lowering state or emit IL.
-Lowerer::ExprType Lowerer::scanExpr(const Expr &e)
-{
-    if (dynamic_cast<const IntExpr *>(&e))
-        return ExprType::I64;
-    if (dynamic_cast<const FloatExpr *>(&e))
-        return ExprType::F64;
-    if (dynamic_cast<const StringExpr *>(&e))
-        return ExprType::Str;
-    if (auto *v = dynamic_cast<const VarExpr *>(&e))
-    {
-        if (!v->name.empty() && v->name.back() == '$')
-            return ExprType::Str;
-        if (!v->name.empty() && v->name.back() == '#')
-            return ExprType::F64;
-        return ExprType::I64;
-    }
-    if (auto *u = dynamic_cast<const UnaryExpr *>(&e))
-        return scanUnaryExpr(*u);
-    if (auto *b = dynamic_cast<const BinaryExpr *>(&e))
-        return scanBinaryExpr(*b);
-    if (auto *arr = dynamic_cast<const ArrayExpr *>(&e))
-        return scanArrayExpr(*arr);
-    if (auto *c = dynamic_cast<const BuiltinCallExpr *>(&e))
-        return scanBuiltinCallExpr(*c);
-    if (auto *c = dynamic_cast<const CallExpr *>(&e))
-    {
-        for (const auto &a : c->args)
-            if (a)
-                scanExpr(*a);
-        return ExprType::I64;
-    }
-    return ExprType::I64;
-}
-
-// Purpose: scan stmt.
-// Parameters: const Stmt &s.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::scanStmt(const Stmt &s)
-{
-    if (auto *l = dynamic_cast<const LetStmt *>(&s))
-    {
-        if (l->expr)
-            scanExpr(*l->expr);
-        if (auto *arr = dynamic_cast<const ArrayExpr *>(l->target.get()))
-            scanExpr(*arr->index);
-    }
-    else if (auto *p = dynamic_cast<const PrintStmt *>(&s))
-    {
-        for (const auto &it : p->items)
-            if (it.kind == PrintItem::Kind::Expr && it.expr)
-                scanExpr(*it.expr);
-    }
-    else if (auto *i = dynamic_cast<const IfStmt *>(&s))
-    {
-        if (i->cond)
-            scanExpr(*i->cond);
-        if (i->then_branch)
-            scanStmt(*i->then_branch);
-        for (const auto &ei : i->elseifs)
-        {
-            if (ei.cond)
-                scanExpr(*ei.cond);
-            if (ei.then_branch)
-                scanStmt(*ei.then_branch);
-        }
-        if (i->else_branch)
-            scanStmt(*i->else_branch);
-    }
-    else if (auto *w = dynamic_cast<const WhileStmt *>(&s))
-    {
-        scanExpr(*w->cond);
-        for (const auto &st : w->body)
-            scanStmt(*st);
-    }
-    else if (auto *f = dynamic_cast<const ForStmt *>(&s))
-    {
-        scanExpr(*f->start);
-        scanExpr(*f->end);
-        if (f->step)
-            scanExpr(*f->step);
-        for (const auto &st : f->body)
-            scanStmt(*st);
-    }
-    else if (auto *inp = dynamic_cast<const InputStmt *>(&s))
-    {
-        needInputLine = true;
-        if (inp->prompt)
-            scanExpr(*inp->prompt);
-        if (inp->var.empty() || inp->var.back() != '$')
-            needRtToInt = true;
-    }
-    else if (auto *d = dynamic_cast<const DimStmt *>(&s))
-    {
-        needAlloc = true;
-        if (d->size)
-            scanExpr(*d->size);
-    }
-    else if (auto *r = dynamic_cast<const RandomizeStmt *>(&s))
-    {
-        trackRuntime(RuntimeFn::RandomizeI64);
-        if (r->seed)
-            scanExpr(*r->seed);
-    }
-    else if (auto *ret = dynamic_cast<const ReturnStmt *>(&s))
-    {
-        if (ret->value)
-            scanExpr(*ret->value);
-    }
-    else if (auto *fn = dynamic_cast<const FunctionDecl *>(&s))
-    {
-        for (const auto &bs : fn->body)
-            scanStmt(*bs);
-    }
-    else if (auto *sub = dynamic_cast<const SubDecl *>(&s))
-    {
-        for (const auto &bs : sub->body)
-            scanStmt(*bs);
-    }
-    else if (auto *lst = dynamic_cast<const StmtList *>(&s))
-    {
-        for (const auto &sub : lst->stmts)
-            scanStmt(*sub);
-    }
-}
-
-// Purpose: scan program.
-// Parameters: const Program &prog.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::scanProgram(const Program &prog)
-{
-    for (const auto &s : prog.procs)
-        scanStmt(*s);
-    for (const auto &s : prog.main)
-        scanStmt(*s);
-}
-
-// Purpose: emit program.
-// Parameters: const Program &prog.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitProgram(const Program &prog)
-{
-    build::IRBuilder &b = *builder;
-
-    std::vector<const Stmt *> mainStmts;
-    for (const auto &s : prog.procs)
-    {
-        if (auto *fn = dynamic_cast<const FunctionDecl *>(s.get()))
-            lowerFunctionDecl(*fn);
-        else if (auto *sub = dynamic_cast<const SubDecl *>(s.get()))
-            lowerSubDecl(*sub);
-    }
-    for (const auto &s : prog.main)
-        mainStmts.push_back(s.get());
-
-    lineBlocks.clear();
-
-    Function &f = b.startFunction("main", Type(Type::Kind::I64), {});
-    func = &f;
-
-    b.addBlock(f, "entry");
-
-    std::vector<int> lines;
-    lines.reserve(mainStmts.size());
-    for (const auto *stmt : mainStmts)
-    {
-        b.addBlock(f, mangler.block("L" + std::to_string(stmt->line)));
-        lines.push_back(stmt->line);
-    }
-    fnExit = f.blocks.size();
-    b.addBlock(f, mangler.block("exit"));
-
-    for (size_t i = 0; i < lines.size(); ++i)
-        lineBlocks[lines[i]] = i + 1;
-
-    vars.clear();
-    arrays.clear();
-    collectVars(mainStmts);
-
-    // allocate slots in entry
-    BasicBlock *entry = &f.blocks.front();
-    cur = entry;
-    for (const auto &v : vars)
-    {
-        curLoc = {};
-        Value slot = emitAlloca(8);
-        varSlots[v] = slot.id; // Value::temp id
-    }
-    if (boundsChecks)
-    {
-        for (const auto &a : arrays)
-        {
-            curLoc = {};
-            Value slot = emitAlloca(8);
-            arrayLenSlots[a] = slot.id;
-        }
-    }
-    if (!mainStmts.empty())
-    {
-        curLoc = {};
-        emitBr(&f.blocks[lineBlocks[mainStmts.front()->line]]);
-    }
-    else
-    {
-        curLoc = {};
-        emitRet(Value::constInt(0));
-    }
-
-    // lower statements sequentially
-    for (size_t i = 0; i < mainStmts.size(); ++i)
-    {
-        cur = &f.blocks[lineBlocks[mainStmts[i]->line]];
-        lowerStmt(*mainStmts[i]);
-        if (!cur->terminated)
-        {
-            BasicBlock *next = (i + 1 < mainStmts.size())
-                                   ? &f.blocks[lineBlocks[mainStmts[i + 1]->line]]
-                                   : &f.blocks[fnExit];
-            curLoc = mainStmts[i]->loc;
-            emitBr(next);
-        }
-    }
-
-    cur = &f.blocks[fnExit];
-    curLoc = {};
-    emitRet(Value::constInt(0));
-}
-
 // Purpose: collect vars.
 // Parameters: const std::vector<const Stmt *> &stmts.
 // Returns: void.
@@ -2415,301 +1874,6 @@ void Lowerer::lowerRandomize(const RandomizeStmt &stmt)
     }
     curLoc = stmt.loc;
     emitCall("rt_randomize_i64", {seed});
-}
-
-// Purpose: lower array addr.
-// Parameters: const ArrayExpr &expr.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::lowerArrayAddr(const ArrayExpr &expr)
-{
-    auto it = varSlots.find(expr.name);
-    assert(it != varSlots.end());
-    Value slot = Value::temp(it->second);
-    Value base = emitLoad(Type(Type::Kind::Ptr), slot);
-    RVal idx = lowerExpr(*expr.index);
-    curLoc = expr.loc;
-    if (boundsChecks)
-    {
-        auto lenIt = arrayLenSlots.find(expr.name);
-        assert(lenIt != arrayLenSlots.end());
-        Value len = emitLoad(Type(Type::Kind::I64), Value::temp(lenIt->second));
-        Value neg = emitBinary(Opcode::SCmpLT, Type(Type::Kind::I1), idx.value, Value::constInt(0));
-        Value ge = emitBinary(Opcode::SCmpGE, Type(Type::Kind::I1), idx.value, len);
-        Value neg64 = emitUnary(Opcode::Zext1, Type(Type::Kind::I64), neg);
-        Value ge64 = emitUnary(Opcode::Zext1, Type(Type::Kind::I64), ge);
-        Value or64 = emitBinary(Opcode::Or, Type(Type::Kind::I64), neg64, ge64);
-        Value cond = emitUnary(Opcode::Trunc1, Type(Type::Kind::I1), or64);
-        size_t curIdx = static_cast<size_t>(cur - &func->blocks[0]);
-        size_t okIdx = func->blocks.size();
-        std::string okLbl = blockNamer ? blockNamer->tag("bc_ok" + std::to_string(boundsCheckId))
-                                       : mangler.block("bc_ok" + std::to_string(boundsCheckId));
-        builder->addBlock(*func, okLbl);
-        size_t failIdx = func->blocks.size();
-        std::string failLbl = blockNamer
-                                  ? blockNamer->tag("bc_fail" + std::to_string(boundsCheckId))
-                                  : mangler.block("bc_fail" + std::to_string(boundsCheckId));
-        builder->addBlock(*func, failLbl);
-        BasicBlock *ok = &func->blocks[okIdx];
-        BasicBlock *fail = &func->blocks[failIdx];
-        cur = &func->blocks[curIdx];
-        ++boundsCheckId;
-        emitCBr(cond, fail, ok);
-        cur = fail;
-        std::string msg = "bounds check failed: " + expr.name + "[i]";
-        Value s = emitConstStr(getStringLabel(msg));
-        emitCall("rt_trap", {s});
-        emitTrap();
-        cur = ok;
-    }
-    Value off = emitBinary(Opcode::Shl, Type(Type::Kind::I64), idx.value, Value::constInt(3));
-    Value ptr = emitBinary(Opcode::GEP, Type(Type::Kind::Ptr), base, off);
-    return ptr;
-}
-
-// Purpose: emit alloca.
-// Parameters: int bytes.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::emitAlloca(int bytes)
-{
-    unsigned id = nextTempId();
-    Instr in;
-    in.result = id;
-    in.op = Opcode::Alloca;
-    in.type = Type(Type::Kind::Ptr);
-    in.operands.push_back(Value::constInt(bytes));
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    return Value::temp(id);
-}
-
-// Purpose: emit load.
-// Parameters: Type ty, Value addr.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::emitLoad(Type ty, Value addr)
-{
-    unsigned id = nextTempId();
-    Instr in;
-    in.result = id;
-    in.op = Opcode::Load;
-    in.type = ty;
-    in.operands.push_back(addr);
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    return Value::temp(id);
-}
-
-// Purpose: emit store.
-// Parameters: Type ty, Value addr, Value val.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitStore(Type ty, Value addr, Value val)
-{
-    Instr in;
-    in.op = Opcode::Store;
-    in.type = ty;
-    in.operands = {addr, val};
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-}
-
-// Purpose: emit for step.
-// Parameters: Value slot, Value step.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitForStep(Value slot, Value step)
-{
-    Value load = emitLoad(Type(Type::Kind::I64), slot);
-    Value add = emitBinary(Opcode::Add, Type(Type::Kind::I64), load, step);
-    emitStore(Type(Type::Kind::I64), slot, add);
-}
-
-// Purpose: emit binary.
-// Parameters: Opcode op, Type ty, Value lhs, Value rhs.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::emitBinary(Opcode op, Type ty, Value lhs, Value rhs)
-{
-    unsigned id = nextTempId();
-    Instr in;
-    in.result = id;
-    in.op = op;
-    in.type = ty;
-    in.operands = {lhs, rhs};
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    return Value::temp(id);
-}
-
-// Purpose: emit unary.
-// Parameters: Opcode op, Type ty, Value val.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::emitUnary(Opcode op, Type ty, Value val)
-{
-    unsigned id = nextTempId();
-    Instr in;
-    in.result = id;
-    in.op = op;
-    in.type = ty;
-    in.operands = {val};
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    return Value::temp(id);
-}
-
-// Purpose: emit br.
-// Parameters: BasicBlock *target.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL. Relies on deterministic block naming via
-// BlockNamer.
-void Lowerer::emitBr(BasicBlock *target)
-{
-    Instr in;
-    in.op = Opcode::Br;
-    in.type = Type(Type::Kind::Void);
-    in.labels.push_back(target->label);
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    cur->terminated = true;
-}
-
-// Purpose: emit cbr.
-// Parameters: Value cond, BasicBlock *t, BasicBlock *f.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL. Relies on deterministic block naming via
-// BlockNamer.
-void Lowerer::emitCBr(Value cond, BasicBlock *t, BasicBlock *f)
-{
-    Instr in;
-    in.op = Opcode::CBr;
-    in.type = Type(Type::Kind::Void);
-    in.operands.push_back(cond);
-    in.labels.push_back(t->label);
-    in.labels.push_back(f->label);
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    cur->terminated = true;
-}
-
-// Purpose: emit call.
-// Parameters: const std::string &callee, const std::vector<Value> &args.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitCall(const std::string &callee, const std::vector<Value> &args)
-{
-    Instr in;
-    in.op = Opcode::Call;
-    in.type = Type(Type::Kind::Void);
-    in.callee = callee;
-    in.operands = args;
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-}
-
-// Purpose: emit call ret.
-// Parameters: Type ty, const std::string &callee, const std::vector<Value> &args.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::emitCallRet(Type ty, const std::string &callee, const std::vector<Value> &args)
-{
-    unsigned id = nextTempId();
-    Instr in;
-    in.result = id;
-    in.op = Opcode::Call;
-    in.type = ty;
-    in.callee = callee;
-    in.operands = args;
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    return Value::temp(id);
-}
-
-// Purpose: emit const str.
-// Parameters: const std::string &globalName.
-// Returns: Value.
-// Side effects: may modify lowering state or emit IL.
-Value Lowerer::emitConstStr(const std::string &globalName)
-{
-    unsigned id = nextTempId();
-    Instr in;
-    in.result = id;
-    in.op = Opcode::ConstStr;
-    in.type = Type(Type::Kind::Str);
-    in.operands.push_back(Value::global(globalName));
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    return Value::temp(id);
-}
-
-// Purpose: emit ret.
-// Parameters: Value v.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitRet(Value v)
-{
-    Instr in;
-    in.op = Opcode::Ret;
-    in.type = Type(Type::Kind::Void);
-    in.operands.push_back(v);
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    cur->terminated = true;
-}
-
-// Purpose: emit ret void.
-// Parameters: none.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitRetVoid()
-{
-    Instr in;
-    in.op = Opcode::Ret;
-    in.type = Type(Type::Kind::Void);
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    cur->terminated = true;
-}
-
-// Purpose: emit trap.
-// Parameters: none.
-// Returns: void.
-// Side effects: may modify lowering state or emit IL.
-void Lowerer::emitTrap()
-{
-    Instr in;
-    in.op = Opcode::Trap;
-    in.type = Type(Type::Kind::Void);
-    in.loc = curLoc;
-    cur->instructions.push_back(in);
-    cur->terminated = true;
-}
-
-// Purpose: get string label.
-// Parameters: const std::string &s.
-// Returns: std::string.
-// Side effects: may modify lowering state or emit IL.
-std::string Lowerer::getStringLabel(const std::string &s)
-{
-    auto it = strings.find(s);
-    if (it != strings.end())
-        return it->second;
-    std::string name = ".L" + std::to_string(strings.size());
-    builder->addGlobalStr(name, s);
-    strings[s] = name;
-    return name;
-}
-
-// Purpose: next temp id.
-// Parameters: none.
-// Returns: unsigned.
-// Side effects: may modify lowering state or emit IL.
-unsigned Lowerer::nextTempId()
-{
-    std::string name = mangler.nextTemp();
-    return static_cast<unsigned>(std::stoul(name.substr(2)));
 }
 
 } // namespace il::frontends::basic
